requirements.txt
flask
flask_restful
newspaper3k